Bug 1540785 - Part 1: Introduce nsDisplayItemBase r=mattwoodrow
***

Differential Revision: https://phabricator.services.mozilla.com/D27579

+++ dom/animation/EffectSet.h
@@ -16,7 +16,7 @@
 #include "nsTHashtable.h"  // For nsTHashtable
 
 class nsPresContext;
-enum class DisplayItemType : uint32_t;
+enum class DisplayItemType : uint8_t;
 
 namespace mozilla {
